+++ src/selftest.c
@@ -164,21 +164,91 @@ static int diag(const char *format, ...)
 }
 
 /*
+ * gotexp_output() - Generate the output used by print_gotexp(). The output is 
+ * returned as an allocated string that must be free()'ed after use. Returns 
+ * NULL if `got` or `exp` is NULL or allocstr() fails. Otherwise, it returns a 
+ * pointer to the string with the output.
+ */
+
+static char *gotexp_output(const char *got, const char *exp)
+{
+	char *s;
+
+	if (!got || !exp)
+		return NULL;
+
+	s = allocstr("         got: '%s'\n"
+	             "    expected: '%s'",
+	             got, exp);
+	if (!s)
+		ok(1, "%s(): allocstr() failed", __func__); /* gncov */
+
+	return s;
+}
+
+/*
  * print_gotexp() - Print the value of the actual and exepected data. Used when 
  * a test fails. Returns 1 if `got` or `exp` is NULL, otherwise 0.
  */
 
 static int print_gotexp(const char *got, const char *exp)
 {
+	char *s;
+
 	if (!got || !exp)
 		return 1;
+	if (!strcmp(got, exp))
+		return 0;
 
-	if (strcmp(got, exp)) {
-		diag("         got: '%s'", got);
-		diag("    expected: '%s'", exp);
-	}
+	s = gotexp_output(got, exp); /* gncov */
+	diag(s); /* gncov */
+	free(s); /* gncov */
 
-	return 0;
+	return 0; /* gncov */
+}
+
+/*
+ * test_gotexp_output() - Tests the gotexp_output() function. print_gotexp() 
+ * can't be tested directly because it would pollute stderr. Returns the number 
+ * of failed tests.
+ */
+
+static int test_gotexp_output(void)
+{
+	int r = 0;
+	char *p, *s;
+
+	diag("Test gotexp_output()");
+
+	r += ok(gotexp_output(NULL, "a") ? 1 : 0,
+	        "gotexp_output(NULL, \"a\")");
+
+	r += ok(strcmp((p = gotexp_output("got this", "expected this")),
+	               "         got: 'got this'\n"
+	               "    expected: 'expected this'") ? 1 : 0,
+	        "gotexp_output(\"got this\", \"expected this\")");
+	free(p);
+
+	r += ok(!print_gotexp(NULL, "expected this"),
+	        "print_gotexp(): Arg is NULL");
+
+	s = "gotexp_output(\"a\", \"a\")";
+	r += ok((p = gotexp_output("a", "a")) ? 0 : 1,
+	        "%s doesn't return NULL", s);
+	r += ok(strcmp(p, "         got: 'a'\n    expected: 'a'") ? 1 : 0,
+	        "%s: Contents is ok", s);
+	free(p);
+
+	s = "gotexp_output() with newline";
+	r += ok((p = gotexp_output("with\nnewline", "also with\nnewline"))
+	        ? 0 : 1,
+	        "%s: Doesn't return NULL", s);
+	r += ok(strcmp(p, "         got: 'with\nnewline'\n"
+	                  "    expected: 'also with\nnewline'") ? 1 : 0,
+	        "%s: Contents is ok", s);
+	free(p);
+
+	return r;
 }
 
 /*
@@ -950,13 +1020,8 @@ static int test_functions(void)
 	diag("Test selftest routines");
 	r += ok(!ok(0, NULL), "ok(0, NULL)");
 	r += test_diag();
+	r += test_gotexp_output();
 	r += test_valgrind_lines();
-
-	diag("Test print_gotexp()");
-	r += ok(print_gotexp("got this", "expected this"),
-	        "print_gotexp() demo");
-	r += ok(!print_gotexp(NULL, "expected this"),
-	        "print_gotexp(): Arg is NULL");
 
 	diag("Test various routines");
 	diag("Test myerror()");